requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-filter==22.1
django-jet-reboot==1.3.3
django-js-asset==2.0.0
django-modeltranslation==0.18.9
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.5.0
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14